Started work on timelogger backend, changed all ids to BIGINT/int64_t

+++ controllers/Scripts.cpp
@@ -18,7 +18,7 @@ void Api::Scripts::getAllScripts(const drogon::HttpRequestPtr&, std::function<vo
 }
 
 void Api::Scripts::getScript(const drogon::HttpRequestPtr&,
-                             std::function<void(const drogon::HttpResponsePtr &)> &&callback, int scriptid) {
+                             std::function<void(const drogon::HttpResponsePtr &)> &&callback, int64_t scriptid) {
     drogon::app().getDbClient()->execSqlAsync("SELECT * FROM scripts WHERE id=$1",
     [callback=std::move(callback)](const drogon::orm::Result& r) {
         Json::Value jsonRow;
@@ -32,7 +32,7 @@ void Api::Scripts::getScript(const drogon::HttpRequestPtr&,
 }
 
 void Api::Scripts::updateScript(const drogon::HttpRequestPtr&,
-                                std::function<void(const drogon::HttpResponsePtr &)> &&callback, int scriptid,
+                                std::function<void(const drogon::HttpResponsePtr &)> &&callback, int64_t scriptid,
                                 std::string &&pCode) {
     std::string codeCopy = pCode;
     drogon::app().getDbClient()->execSqlAsync("UPDATE SCRIPTS SET code=$1 WHERE id=$2 RETURNING name",
@@ -49,7 +49,7 @@ void Api::Scripts::updateScript(const drogon::HttpRequestPtr&,
 }
 
 void Api::Scripts::runScript(const drogon::HttpRequestPtr&,
-                             std::function<void(const drogon::HttpResponsePtr &)> &&callback, int scriptid, std::string&& pParam) {
+                             std::function<void(const drogon::HttpResponsePtr &)> &&callback, int64_t scriptid, std::string&& pParam) {
     drogon::app().getDbClient()->execSqlAsync("SELECT name FROM scripts WHERE id=$1",
     [callback=std::move(callback),param=std::move(pParam)](const drogon::orm::Result& r) {
         auto name = r.at(0)["name"].as<std::string>();
@@ -86,7 +86,7 @@ void Api::Scripts::runScript(const drogon::HttpRequestPtr&,
 }
 
 void Api::Scripts::drawScript(const drogon::HttpRequestPtr&,
-                              std::function<void(const drogon::HttpResponsePtr &)> &&callback, int scriptid) {
+                              std::function<void(const drogon::HttpResponsePtr &)> &&callback, int64_t scriptid) {
     drogon::app().getDbClient()->execSqlAsync("SELECT name FROM scripts WHERE id=$1",
     [callback=std::move(callback)](const drogon::orm::Result& r) {
         auto name = r.at(0)["name"].as<std::string>();
@@ -111,7 +111,7 @@ void Api::Scripts::drawScript(const drogon::HttpRequestPtr&,
 }
 
 void Api::Scripts::deleteScript(const drogon::HttpRequestPtr&,
-                                std::function<void(const drogon::HttpResponsePtr &)> &&callback, int scriptid) {
+                                std::function<void(const drogon::HttpResponsePtr &)> &&callback, int64_t scriptid) {
     drogon::app().getDbClient()->execSqlAsync("DELETE FROM scripts WHERE id=$1 RETURNING name",
     [callback=std::move(callback)](const drogon::orm::Result& r) {
         auto name = r.at(0)["name"].as<std::string>();
